PPAD-42: separate account-details from account js
